Option to keep submit errors on change (#1735)
* Add clearErrorsOnChange config property to reduxForm() (#1708)

* Make clearErrorsOnChange also affect global errors (#1708)

* Fix clearing old global errors on stopSubmit (#1708)

* Global errors were always not cleared by stopSubmit
* Clearing field submit errors was not tested

* Rename (and invert) clearErrorsOnChange to persistentSubmitErrors (#1708)

* Fix submit not working when persistentSubmitErrors is enabled and there are old submit errors (#1708)

* Fix lint errors

+++ src/reducer.js
@@ -146,7 +146,7 @@ const createReducer = structure => {
       }
       return result
     },
-    [CHANGE](state, { meta: { field, touch }, payload }) {
+    [CHANGE](state, { meta: { field, touch, persistentSubmitErrors }, payload }) {
       let result = state
       const initial = getIn(result, `initial.${field}`)
       if (initial === undefined && payload === '') {
@@ -155,9 +155,13 @@ const createReducer = structure => {
         result = setIn(result, `values.${field}`, payload)
       }
       result = deleteInWithCleanUp(result, `asyncErrors.${field}`)
-      result = deleteInWithCleanUp(result, `submitErrors.${field}`)
+      if (!persistentSubmitErrors) {
+        result = deleteInWithCleanUp(result, `submitErrors.${field}`)
+      }
       result = deleteInWithCleanUp(result, `fields.${field}.autofilled`)
-      result = deleteInWithCleanUp(result, 'error')
+      if (!persistentSubmitErrors) {
+        result = deleteInWithCleanUp(result, 'error')
+      }
       if (touch) {
         result = setIn(result, `fields.${field}.touched`, true)
         result = setIn(result, 'anyTouched', true)
@@ -269,6 +273,8 @@ const createReducer = structure => {
         const { _error, ...fieldErrors } = payload
         if (_error) {
           result = setIn(result, 'error', _error)
+        } else {
+          result = deleteIn(result, 'error')
         }
         if (Object.keys(fieldErrors).length) {
           result = setIn(result, 'submitErrors', fromJS(fieldErrors))
